Task 3 - Resume Screening/requirements.txt
langchain
langchain-groq
python-dotenv
langsmith
